TRACK-466 Split AccessionsNamespace
Introduce a new class `SearchFieldNamespaces` that constructs the namespaces and
exposes them to the application as needed. For now, it only constructs the
existing accessions namespace.

Split `AccessionsNamespace` into a series of plain (non-`@ManagedBean`) classes.

No change in functionality here. A subsequent change will start introducing more
namespaces to the hierarchy.
