+++ src/main/kotlin/com/terraformation/backend/search/NestedQueryBuilder.kt
@@ -1,7 +1,7 @@
 package com.terraformation.backend.search
 
 import com.terraformation.backend.search.field.SearchField
-import com.terraformation.backend.seedbank.search.AccessionsNamespace
+import com.terraformation.backend.search.namespace.AccessionsNamespace
 import com.terraformation.backend.util.MemoizedValue
 import org.jooq.Condition
 import org.jooq.DSLContext
